Add basic tests for PostAnalyzer
