+++ src/PostAnalyzer.cs
@@ -39,26 +39,6 @@ using System.Text.RegularExpressions;
 namespace Eleia
 {
     /// <summary>
-    /// Represents any possible problem with a post, like unformatted code,
-    /// bad title, bad tags and so on
-    /// </summary>
-    public abstract class PostProblems
-    {
-        public float Probability { get; set; }
-    }
-
-    /// <summary>
-    /// Represents that a class has a not formatted code somewhere
-    /// </summary>
-    public class NotFormattedCodeFound : PostProblems
-    {
-        public override string ToString()
-        {
-            return $"Potentially not formatted code found (prob: {Probability})";
-        }
-    }
-
-    /// <summary>
     /// Analyzes the post on the Coyote forum in search of problems
     /// </summary>
     public class PostAnalyzer
